refactor(swiftui): share the colorset helpers between asset generators
The upcoming themed-colours generator needs the same colorset-writing
helpers as the semantic generator, without folding themed hues into
the semantic shorthand namespaces. Move ColorValue, ColorResource,
parseThemeColors, generateColorAsset, formatColorComponent,
sanitizeGroup and sanitizeSwiftName into the shared loader (which has
no main() and so can be imported), and add a lemonadeAssetName helper
so both generators build asset names the same way.

No-op refactor: regenerated output is byte-identical to before.
