+++ scripts/swiftui-color-assets-generator.main.kts
@@ -19,21 +19,6 @@ import kotlin.math.roundToInt
  * Usage: kotlin scripts/swiftui-color-assets-generator.main.kts
  */
 
-data class ColorValue(
-    val r: Double,
-    val g: Double,
-    val b: Double,
-    val a: Double,
-)
-
-data class ColorResource(
-    val group: String,      // e.g., "Content", "Background", "Border", "Interaction"
-    val name: String,       // e.g., "contentPrimary", "bgDefault"
-    val assetName: String,  // e.g., "lemonade-content-primary"
-    val lightColor: ColorValue,
-    val darkColor: ColorValue?, // null if dark theme not available yet
-)
-
 fun main() {
     val themeFile = tokenFile("theme-colors.light.tokens.json")
 
@@ -74,7 +59,7 @@ fun main() {
             val parts = key.split("/")
             val group = parts.getOrNull(0)?.sanitizeGroup() ?: "Other"
             val name = parts.last().sanitizeSwiftName()
-            val assetName = "lemonade-${parts.joinToString("-") { it.lowercase().replace("_", "-") }}"
+            val assetName = lemonadeAssetName(key)
 
             ColorResource(
                 group = group,
@@ -104,74 +89,6 @@ fun main() {
         error.printStackTrace()
         throw error
     }
-}
-
-fun parseThemeColors(files: List<File>, modeName: String): Map<String, ColorValue> {
-    val colors = linkedMapOf<String, ColorValue>()
-    readFileResourceFileByModeRaw(files, modeName) { name, resolved ->
-        colors[name] = ColorValue(
-            r = resolved.getDouble("r"),
-            g = resolved.getDouble("g"),
-            b = resolved.getDouble("b"),
-            a = resolved.optDouble("a", 1.0),
-        )
-    }
-    return colors
-}
-
-fun generateColorAsset(assetsDir: File, resource: ColorResource) {
-    val colorsetDir = File(assetsDir, "${resource.assetName}.colorset")
-    colorsetDir.mkdirs()
-
-    val light = resource.lightColor
-    val dark = resource.darkColor ?: resource.lightColor // Fallback to light if no dark
-
-    val contentsJson = """
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "${formatColorComponent(light.a)}",
-          "blue" : "${formatColorComponent(light.b)}",
-          "green" : "${formatColorComponent(light.g)}",
-          "red" : "${formatColorComponent(light.r)}"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "${formatColorComponent(dark.a)}",
-          "blue" : "${formatColorComponent(dark.b)}",
-          "green" : "${formatColorComponent(dark.g)}",
-          "red" : "${formatColorComponent(dark.r)}"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
-""".trimIndent()
-
-    File(colorsetDir, "Contents.json").writeText(contentsJson)
-}
-
-fun formatColorComponent(value: Double): String {
-    return "%.3f".format(Locale.US, value)
 }
 
 fun generateColorShorthand(resources: List<ColorResource>): String {
@@ -253,22 +170,6 @@ fun generateColorShorthand(resources: List<ColorResource>): String {
 
         appendLine("}")
     }
-}
-
-fun String.sanitizeGroup(): String {
-    return split("/").firstOrNull()
-        ?.split("-")
-        ?.joinToString("") { it.replaceFirstChar { c -> c.uppercase() } }
-        ?: "Other"
-}
-
-fun String.sanitizeSwiftName(): String {
-    return split("-")
-        .mapIndexed { index, word ->
-            if (index == 0) word.lowercase()
-            else word.replaceFirstChar { it.uppercase() }
-        }
-        .joinToString("")
 }
 
 main()
