+++ scripts/swiftui-resource-file-loading.main.kts
@@ -3,6 +3,8 @@
 
 import org.json.JSONObject
 import java.io.File
+import java.util.Locale
+import kotlin.math.roundToInt
 
 data class ResourceData<T>(
     val groups: List<String>,
@@ -360,4 +362,120 @@ fun readFileResourceFileByModeRaw(
         return
     }
     error("No token file provides mode '$modeName'")
+}
+
+data class ColorValue(
+    val r: Double,
+    val g: Double,
+    val b: Double,
+    val a: Double,
+)
+
+data class ColorResource(
+    val group: String,      // e.g., "Content", "Background", "Border", "Interaction"
+    val name: String,       // e.g., "contentPrimary", "bgDefault"
+    val assetName: String,  // e.g., "lemonade-content-primary"
+    val lightColor: ColorValue,
+    val darkColor: ColorValue?, // null if dark theme not available yet
+)
+
+fun parseThemeColors(files: List<File>, modeName: String): Map<String, ColorValue> {
+    val colors = linkedMapOf<String, ColorValue>()
+    readFileResourceFileByModeRaw(files, modeName) { name, resolved ->
+        colors[name] = ColorValue(
+            r = resolved.getDouble("r"),
+            g = resolved.getDouble("g"),
+            b = resolved.getDouble("b"),
+            a = resolved.optDouble("a", 1.0),
+        )
+    }
+    return colors
+}
+
+fun generateColorAsset(assetsDir: File, resource: ColorResource) {
+    val colorsetDir = File(assetsDir, "${resource.assetName}.colorset")
+    colorsetDir.mkdirs()
+
+    val light = resource.lightColor
+    val dark = resource.darkColor ?: resource.lightColor // Fallback to light if no dark
+
+    val contentsJson = """
+{
+  "colors" : [
+    {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "${formatColorComponent(light.a)}",
+          "blue" : "${formatColorComponent(light.b)}",
+          "green" : "${formatColorComponent(light.g)}",
+          "red" : "${formatColorComponent(light.r)}"
+        }
+      },
+      "idiom" : "universal"
+    },
+    {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "${formatColorComponent(dark.a)}",
+          "blue" : "${formatColorComponent(dark.b)}",
+          "green" : "${formatColorComponent(dark.g)}",
+          "red" : "${formatColorComponent(dark.r)}"
+        }
+      },
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
+""".trimIndent()
+
+    File(colorsetDir, "Contents.json").writeText(contentsJson)
+}
+
+fun formatColorComponent(value: Double): String {
+    return "%.3f".format(Locale.US, value)
+}
+
+fun String.sanitizeGroup(): String {
+    return split("/").firstOrNull()
+        ?.split("-")
+        ?.joinToString("") { it.replaceFirstChar { c -> c.uppercase() } }
+        ?: "Other"
+}
+
+fun String.sanitizeSwiftName(): String {
+    return split("-")
+        .mapIndexed { index, word ->
+            if (index == 0) word.lowercase()
+            else word.replaceFirstChar { it.uppercase() }
+        }
+        .joinToString("")
+}
+
+/**
+ * The asset-catalog name for a token path, e.g. `Background/bg-blue` with prefix
+ * `themed` becomes `lemonade-themed-background-bg-blue`.
+ *
+ * Both the colorset folders and the `Color("...")` lookups must be built from
+ * this one function. PR #361 hit the alternative: two converters constructing
+ * the name separately, only one of them transforming it, and the lookups
+ * pointing at folders that did not exist — which fails silently at runtime
+ * rather than failing the build.
+ */
+fun lemonadeAssetName(tokenPath: String, prefix: String? = null): String {
+    val slug = tokenPath.split("/").joinToString("-") { part ->
+        part.lowercase().replace("_", "-")
+    }
+    return if (prefix == null) "lemonade-$slug" else "lemonade-$prefix-$slug"
 }
